More improvements to the API are possible as per discussed in Discord, for example making renderers implement a common trait so that they don't own/repeat the generic logic loop - until then, this will do.

+++ src/render/wgpu/pipeline.rs
@@ -202,7 +202,7 @@ pub struct InoxPipeline {
     pub uniform_layout: BindGroupLayout,
     pub texture_layout: BindGroupLayout,
     pub texture_format: TextureFormat,
-    pub uniform_alignment_needed: u64,
+    pub uniform_alignment_needed: usize,
 }
 
 impl InoxPipeline {
@@ -326,9 +326,9 @@ impl InoxPipeline {
             uniform_layout,
             texture_layout,
             texture_format,
-            uniform_alignment_needed: Uniform::min_size()
-                .get()
-                .max(min_uniform_buffer_offset_alignment.into()),
+            uniform_alignment_needed: (Uniform::min_size().get())
+                .max(min_uniform_buffer_offset_alignment.into())
+                as usize,
         }
     }
 }
